feat: Added support for Mongo v5+
