+++ lib/instrumentation/mongodb/common.js
@@ -72,9 +72,13 @@ common.instrumentBulkOperation = function instrumentBulkOperation(shim, BulkOper
 common.instrumentDb = function instrumentDb(shim, Db) {
   if (Db && Db.prototype) {
     const proto = Db.prototype
-    shim.recordOperation(proto, DB_OPS, new OperationSpec({ callback: shim.LAST, opaque: true }))
+    shim.recordOperation(
+      proto,
+      DB_OPS,
+      new OperationSpec({ callback: shim.LAST, opaque: true, promise: true })
+    )
     // link to client.connect(removed in v4.0)
-    shim.recordOperation(Db, 'connect', new OperationSpec({ callback: shim.LAST }))
+    shim.recordOperation(Db, 'connect', new OperationSpec({ callback: shim.LAST, promise: true }))
   }
 }
 
@@ -133,7 +137,7 @@ common.makeBulkDescFunc = function makeBulkDescFunc(shim) {
  *
  * @param {Shim} shim instance of shim
  * @param {object} instrumenter instance of mongo APM class
- * @param {object} [options={}] provide command names to skip updating host/port as they are unrelated to the active query.  This is only in v3 because after every command is runs `endSessions` which runs on the admin database
+ * @param {object} [options] provide command names to skip updating host/port as they are unrelated to the active query.  This is only in v3 because after every command is runs `endSessions` which runs on the admin database
  */
 common.captureAttributesOnStarted = function captureAttributesOnStarted(
   shim,
@@ -214,7 +218,6 @@ function setHostPort(shim, connStr, db, client) {
  */
 function getInstanceAttributeParameters(shim, mongo) {
   let params
-
   if (mongo?.s?.topology) {
     shim.logger.trace('Adding datastore instance attributes from mongo.s.db + mongo.s.topology')
     const databaseName = mongo?.s?.db?.databaseName || mongo?.s?.namespace?.db || null
@@ -224,6 +227,10 @@ function getInstanceAttributeParameters(shim, mongo) {
     const databaseName = mongo?.s?.db?.databaseName || null
     const hosts = mongo.s.db.s.client.s.options.hosts
     params = getParametersFromHosts(hosts, databaseName)
+  } else if (mongo?.s?.db?.client?.topology) {
+    const databaseName = mongo?.s?.namespace?.db
+    const topology = mongo.s.db.client.topology
+    params = getParametersFromTopology(topology, databaseName)
   } else {
     shim.logger.trace('Could not find datastore instance attributes.')
     params = new DatastoreParameters()
@@ -272,6 +279,11 @@ function getParametersFromTopology(conf, database) {
     ;[{ host, port }] = conf.s.options.servers
   }
 
+  // hosts is an array but we will always pull the first for consistency
+  if (conf?.s?.options?.hosts?.length) {
+    ;[{ host, port }] = conf.s.options.hosts
+  }
+
   // host is a domain socket. set host as localhost and use the domain
   // socket host as the port
   if (host && host.endsWith('.sock')) {
@@ -280,7 +292,7 @@ function getParametersFromTopology(conf, database) {
   }
 
   return new DatastoreParameters({
-    host: host,
+    host,
     port_path_or_id: port,
     database_name: database
   })
