+++ test/versioned/mongodb/collection-update.tap.js
@@ -19,53 +19,122 @@ const { pkgVersion, STATEMENT_PREFIX } = require('./common')
  * @param {Number} params.count, optional
  * @param {string} params.keyPrefix prefix where the count exists
  * @param {Object} params.extraValues extra fields to assert on >=4.0.0 version of module
- * @param {Object} params.legaycValues extra fields to assert on <4.0.0 version of module
  */
-function assertExpectedResult({ t, data, count, keyPrefix, extraValues, legacyValues }) {
-  if (semver.satisfies(pkgVersion, '<4')) {
-    const expectedResult = { ok: 1, ...legacyValues }
-    if (count) {
-      expectedResult.n = count
-    }
-    t.same(data.result, expectedResult)
-  } else {
-    const expectedResult = { acknowledged: true, ...extraValues }
-    if (count) {
-      expectedResult[`${keyPrefix}Count`] = count
-    }
-    t.same(data, expectedResult)
+function assertExpectedResult({ t, data, count, keyPrefix, extraValues }) {
+  const expectedResult = { acknowledged: true, ...extraValues }
+  if (count) {
+    expectedResult[`${keyPrefix}Count`] = count
   }
+  t.same(data, expectedResult)
 }
 
-common.test('deleteMany', function deleteManyTest(t, collection, verify) {
-  collection.deleteMany({ mod10: 5 }, function done(err, data) {
-    t.error(err)
-    assertExpectedResult({
-      t,
-      data,
-      count: 3,
-      keyPrefix: 'deleted'
-    })
-    verify(null, [`${STATEMENT_PREFIX}/deleteMany`, 'Callback: done'], ['deleteMany'])
+common.test('deleteMany', async function deleteManyTest(t, collection, verify) {
+  const data = await collection.deleteMany({ mod10: 5 })
+  assertExpectedResult({
+    t,
+    data,
+    count: 3,
+    keyPrefix: 'deleted'
   })
+  verify(null, [`${STATEMENT_PREFIX}/deleteMany`], ['deleteMany'], { strict: false })
 })
 
-common.test('deleteOne', function deleteOneTest(t, collection, verify) {
-  collection.deleteOne({ mod10: 5 }, function done(err, data) {
-    t.error(err)
-    assertExpectedResult({
-      t,
-      data,
-      count: 1,
-      keyPrefix: 'deleted'
-    })
-    verify(null, [`${STATEMENT_PREFIX}/deleteOne`, 'Callback: done'], ['deleteOne'])
+common.test('deleteOne', async function deleteOneTest(t, collection, verify) {
+  const data = await collection.deleteOne({ mod10: 5 })
+  assertExpectedResult({
+    t,
+    data,
+    count: 1,
+    keyPrefix: 'deleted'
   })
+  verify(null, [`${STATEMENT_PREFIX}/deleteOne`], ['deleteOne'], { strict: false })
 })
 
-common.test('insert', function insertTest(t, collection, verify) {
-  collection.insert({ foo: 'bar' }, function done(err, data) {
-    t.error(err)
+common.test('insertMany', async function insertManyTest(t, collection, verify) {
+  const data = await collection.insertMany([{ foo: 'bar' }, { foo: 'bar2' }])
+  assertExpectedResult({
+    t,
+    data,
+    count: 2,
+    keyPrefix: 'inserted',
+    extraValues: {
+      insertedIds: {
+        0: data.insertedIds[0],
+        1: data.insertedIds[1]
+      }
+    }
+  })
+
+  verify(null, [`${STATEMENT_PREFIX}/insertMany`], ['insertMany'], { strict: false })
+})
+
+common.test('insertOne', async function insertOneTest(t, collection, verify) {
+  const data = await collection.insertOne({ foo: 'bar' })
+  assertExpectedResult({
+    t,
+    data,
+    extraValues: {
+      insertedId: data.insertedId
+    }
+  })
+
+  verify(null, [`${STATEMENT_PREFIX}/insertOne`], ['insertOne'], { strict: false })
+})
+
+common.test('replaceOne', async function replaceOneTest(t, collection, verify) {
+  const data = await collection.replaceOne({ i: 5 }, { foo: 'bar' })
+  assertExpectedResult({
+    t,
+    data,
+    count: 1,
+    keyPrefix: 'modified',
+    extraValues: {
+      matchedCount: 1,
+      upsertedCount: 0,
+      upsertedId: null
+    }
+  })
+
+  verify(null, [`${STATEMENT_PREFIX}/replaceOne`], ['replaceOne'], { strict: false })
+})
+
+common.test('updateMany', async function updateManyTest(t, collection, verify) {
+  const data = await collection.updateMany({ mod10: 5 }, { $set: { a: 5 } })
+  assertExpectedResult({
+    t,
+    data,
+    count: 3,
+    keyPrefix: 'modified',
+    extraValues: {
+      matchedCount: 3,
+      upsertedCount: 0,
+      upsertedId: null
+    }
+  })
+
+  verify(null, [`${STATEMENT_PREFIX}/updateMany`], ['updateMany'], { strict: false })
+})
+
+common.test('updateOne', async function updateOneTest(t, collection, verify) {
+  const data = await collection.updateOne({ i: 5 }, { $set: { a: 5 } })
+  assertExpectedResult({
+    t,
+    data,
+    count: 1,
+    keyPrefix: 'modified',
+    extraValues: {
+      matchedCount: 1,
+      upsertedCount: 0,
+      upsertedId: null
+    }
+  })
+
+  verify(null, [`${STATEMENT_PREFIX}/updateOne`], ['updateOne'], { strict: false })
+})
+
+if (semver.satisfies(pkgVersion, '<5.0.0')) {
+  common.test('insert', async function insertTest(t, collection, verify) {
+    const data = await collection.insert({ foo: 'bar' })
     assertExpectedResult({
       t,
       data,
@@ -78,51 +147,11 @@ common.test('insert', function insertTest(t, collection, verify) {
       }
     })
 
-    verify(null, [`${STATEMENT_PREFIX}/insert`, 'Callback: done'], ['insert'])
+    verify(null, [`${STATEMENT_PREFIX}/insert`], ['insert'], { strict: false })
   })
-})
 
-common.test('insertMany', function insertManyTest(t, collection, verify) {
-  collection.insertMany([{ foo: 'bar' }, { foo: 'bar2' }], function done(err, data) {
-    t.error(err)
-    assertExpectedResult({
-      t,
-      data,
-      count: 2,
-      keyPrefix: 'inserted',
-      extraValues: {
-        insertedIds: {
-          0: {},
-          1: {}
-        }
-      }
-    })
-
-    verify(null, [`${STATEMENT_PREFIX}/insertMany`, 'Callback: done'], ['insertMany'])
-  })
-})
-
-common.test('insertOne', function insertOneTest(t, collection, verify) {
-  collection.insertOne({ foo: 'bar' }, function done(err, data) {
-    t.error(err)
-    assertExpectedResult({
-      t,
-      data,
-      legacyValues: {
-        n: 1
-      },
-      extraValues: {
-        insertedId: {}
-      }
-    })
-
-    verify(null, [`${STATEMENT_PREFIX}/insertOne`, 'Callback: done'], ['insertOne'])
-  })
-})
-
-common.test('remove', function removeTest(t, collection, verify) {
-  collection.remove({ mod10: 5 }, function done(err, data) {
-    t.error(err)
+  common.test('remove', async function removeTest(t, collection, verify) {
+    const data = await collection.remove({ mod10: 5 })
     assertExpectedResult({
       t,
       data,
@@ -130,21 +159,16 @@ common.test('remove', function removeTest(t, collection, verify) {
       keyPrefix: 'deleted'
     })
 
-    verify(null, [`${STATEMENT_PREFIX}/remove`, 'Callback: done'], ['remove'])
+    verify(null, [`${STATEMENT_PREFIX}/remove`], ['remove'], { strict: false })
   })
-})
 
-common.test('replaceOne', function replaceOneTest(t, collection, verify) {
-  collection.replaceOne({ i: 5 }, { foo: 'bar' }, function done(err, data) {
-    t.error(err)
+  common.test('update', async function updateTest(t, collection, verify) {
+    const data = await collection.update({ i: 5 }, { $set: { foo: 'bar' } })
     assertExpectedResult({
       t,
       data,
       count: 1,
       keyPrefix: 'modified',
-      legacyValues: {
-        nModified: 1
-      },
       extraValues: {
         matchedCount: 1,
         upsertedCount: 0,
@@ -152,83 +176,6 @@ common.test('replaceOne', function replaceOneTest(t, collection, verify) {
       }
     })
 
-    verify(null, [`${STATEMENT_PREFIX}/replaceOne`, 'Callback: done'], ['replaceOne'])
-  })
-})
-
-if (semver.satisfies(pkgVersion, '<4')) {
-  common.test('save', function saveTest(t, collection, verify) {
-    collection.save({ foo: 'bar' }, function done(err, data) {
-      t.error(err)
-      t.same(data.result, { ok: 1, n: 1 })
-
-      verify(null, [`${STATEMENT_PREFIX}/save`, 'Callback: done'], ['save'])
-    })
+    verify(null, [`${STATEMENT_PREFIX}/update`], ['update'], { strict: false })
   })
 }
-
-common.test('update', function updateTest(t, collection, verify) {
-  collection.update({ i: 5 }, { $set: { foo: 'bar' } }, function done(err, data) {
-    t.error(err)
-    assertExpectedResult({
-      t,
-      data,
-      count: 1,
-      keyPrefix: 'modified',
-      legacyValues: {
-        nModified: 1
-      },
-      extraValues: {
-        matchedCount: 1,
-        upsertedCount: 0,
-        upsertedId: null
-      }
-    })
-
-    verify(null, [`${STATEMENT_PREFIX}/update`, 'Callback: done'], ['update'])
-  })
-})
-
-common.test('updateMany', function updateManyTest(t, collection, verify) {
-  collection.updateMany({ mod10: 5 }, { $set: { a: 5 } }, function done(err, data) {
-    t.error(err)
-    assertExpectedResult({
-      t,
-      data,
-      count: 3,
-      keyPrefix: 'modified',
-      legacyValues: {
-        nModified: 3
-      },
-      extraValues: {
-        matchedCount: 3,
-        upsertedCount: 0,
-        upsertedId: null
-      }
-    })
-
-    verify(null, [`${STATEMENT_PREFIX}/updateMany`, 'Callback: done'], ['updateMany'])
-  })
-})
-
-common.test('updateOne', function updateOneTest(t, collection, verify) {
-  collection.updateOne({ i: 5 }, { $set: { a: 5 } }, function done(err, data) {
-    t.notOk(err, 'should not error')
-    assertExpectedResult({
-      t,
-      data,
-      count: 1,
-      keyPrefix: 'modified',
-      legacyValues: {
-        nModified: 1
-      },
-      extraValues: {
-        matchedCount: 1,
-        upsertedCount: 0,
-        upsertedId: null
-      }
-    })
-
-    verify(null, [`${STATEMENT_PREFIX}/updateOne`, 'Callback: done'], ['updateOne'])
-  })
-})
