+++ test/versioned/mongodb/package.json
@@ -8,10 +8,29 @@
         "node": ">=16"
       },
       "dependencies": {
-        "mongodb": ">=2.1 < 4.0.0 || >= 4.1.4 < 5"
+        "mongodb": {
+          "versions": ">=2.1 < 4.0.0",
+          "samples": "2"
+        }
       },
       "files": [
-        "bulk.tap.js",
+        "legacy/bulk.tap.js",
+        "legacy/cursor.tap.js",
+        "legacy/db.tap.js",
+        "legacy/find.tap.js",
+        "legacy/index.tap.js",
+        "legacy/misc.tap.js",
+        "legacy/update.tap.js"
+      ]
+    },
+    {
+      "engines": {
+        "node": ">=16"
+      },
+      "dependencies": {
+        "mongodb": ">=4.1.4"
+      },
+      "files": [
         "collection-find.tap.js",
         "collection-index.tap.js",
         "collection-misc.tap.js",
